requirements.txt
# Sell Option Scanner — Dependencies
# =====================================
# Install with:
#   Mac/Linux:  pip3 install -r requirements.txt
#   Windows:    py -3.11 -m pip install -r requirements.txt

# Core math
scipy>=1.11.0           # Black-Scholes normal distribution
numpy>=1.24.0           # Array operations

# Option chain data (free, no login — replaces IBKR for scanning)
yfinance>=0.2.40        # Yahoo Finance option chains + price history
pandas>=2.0.0           # Required by yfinance for option chain DataFrames

# IBKR API (only needed if set data_source ibkr)
ib_insync>=0.9.86       # TWS/IB Gateway API wrapper

# Config / environment
python-dotenv>=1.0.0    # Load .env credentials file

# Database
supabase>=2.0.0         # Supabase client (trade_candidates, trade_log)

# Timezone (improves DST accuracy for ET market-hours checks)
pytz>=2023.3

# Web framework (SaaS backend)
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Auth (JWT verification)
python-jose[cryptography]>=3.3.0
httpx>=0.27.0

# Testing
pytest>=7.0.0
